+++ durihana/components/schedule/UserCalendar.tsx
@@ -20,12 +20,6 @@ export default function UserCalendar({ userId }: ScheduleProps) {
     reservationScheduleDates,
   } = useUserCalendar(userId);
 
-  // // 계좌 생성 후 달력 새로고침을 위한 함수
-  // const handleAccountCreated = () => {
-  //   // 현재 월의 일정 다시 로드하기 위해 상태를 강제로 업데이트
-  //   setCalendarMonth(calendarMonth);
-  // };
-
   return (
     <div className='bg-background mx-auto flex h-screen w-full max-w-md flex-col'>
       {/* 달력 - 고정 영역 */}
